backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.8.3
numpy==1.26.2
Pillow==10.1.0
python-dotenv==1.0.0
scikit-learn==1.3.2
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0
pinecone-client==3.2.2
tensorflow==2.15.0
tensorflow-hub==0.15.0
gunicorn==21.2.0
